Added data for equipment, inventory, bank, kod, guidepcs, and more.
