--- src/main/java/com/larsvansoest/runelite/modules/itemwidgets/ItemWidgetID.java
@@ -1,16 +1,0 @@
-package com.larsvansoest.runelite.modules.itemwidgets;
-
-class ItemWidgetID
-{
-	public static int EQUIPMENT_ITEMS_GROUP_ID = 84;
-
-	public static class Shop
-	{
-		public static int ITEMS_CONTAINER = 16;
-	}
-
-	public static class Equipment
-	{
-		public static int ITEM_SLOTS = 1;
-	}
-}
